pyproject.toml
[project]
name = "repo1-mcp-server"
version = "0.1.0"
description = "MCP-compliant server exposing financial data tools for LLM tool-calling"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "mcp>=1.1.0",
    "faker>=28.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "black>=24.4.0",
    "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
include = ["app*", "scripts*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

# ── Ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ── Black ─────────────────────────────────────────────────────────────────────
[tool.black]
target-version = ["py311"]
line-length = 100

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
